lambda_3/requirements.txt
s3fs
psycopg2-binary
datetime
python-dotenv
